+++ lib/get_core/src/flutter_engine.dart
@@ -1,6 +1,10 @@
 import 'package:flutter/widgets.dart';
 
+/// A utility class for accessing the engine instance.
 class Engine {
+  /// Returns the instance of the widgets binding.
+  ///
+  /// This method ensures that the widgets binding is initialized and returns the instance.
   static WidgetsBinding get instance {
     return WidgetsFlutterBinding.ensureInitialized();
   }